ver5.1

+++ APE.cpp
@@ -11,7 +11,7 @@
 using namespace cv;
 using namespace std;
 
-void APE(pose *p, parameter *para, const Mat &marker, const Mat &img, gpu::GpuMat &marker_d, const float &Sfx, const float &Sfy, const int &Px, const int &Py, 
+void APE(pose *p, parameter *para, const Mat &marker, const Mat &img, gpu::GpuMat &marker_d, const float &Sfx, const float &Sfy, const float &Px, const float &Py, 
          const float &minDim, const float &tzMin, const float &tzMax, const float &delta, const bool photo, const bool verbose) {
   gpu::setDevice(0);
